p2p: smaller max tx size
All L2 transactions must be able to be included
in an L1 block as part of a L1 transaction.
It would be very inconvenient if users submitted
transactions to L2 that resulted in the L1 transaction
not being able to be gossiped around the L1 p2p network.
This commit restricts the max size of a L2 transaction
being accepted into the mempool such that it can be
included in an L1 transaction.