core: Add version to DepositTx (#19)
This adds a second byte after the EIP-2718 Type byte to deposit
transactions that versions the deposit. It is placed prior to the start
of the RLP to make it easier to extend the transaction in the future. We
choose to version inside the EIP-2718 envelope to limit the number of
EIP-2718 prefixes we take for deposit transactions.